+++ client/app/deskaas/deskaas-details/deskaas-details.js
@@ -1,17 +1,16 @@
 "use strict";
 
 angular.module("managerApp")
-.config(function ($stateProvider) {
+    .config($stateProvider => {
 
-    $stateProvider.state("deskaas.details", {
-        url          : "/:serviceName?action&token",
-        templateUrl  : "app/deskaas/deskaas-details/deskaas-details.html",
-        controller   : "DeskaasDetailsCtrl",
-        controllerAs : "DeskaasDetailsCtrl",
-        translations : ["common", "deskaas", "deskaas/deskaas-details"],
-        params: {
-            followTask: null
-        }
+        $stateProvider.state("deskaas.details", {
+            url: "/:serviceName?action&token",
+            templateUrl: "app/deskaas/deskaas-details/deskaas-details.html",
+            controller: "DeskaasDetailsCtrl",
+            controllerAs: "$ctrl",
+            translations: ["common", "deskaas", "deskaas/deskaas-details"],
+            params: {
+                followTask: null
+            }
+        });
     });
-
-});