+++ client/app/iplb/home/updateQuota/iplb-update-quota.controller.js
@@ -1,7 +1,8 @@
 class IpLoadBalancerUpdateQuotaCtrl {
-    constructor ($stateParams, $uibModalInstance, IpLoadBalancerHomeService, quota) {
+    constructor ($stateParams, $uibModalInstance, ControllerHelper, IpLoadBalancerHomeService, quota) {
         this.$stateParams = $stateParams;
         this.$uibModalInstance = $uibModalInstance;
+        this.ControllerHelper = ControllerHelper;
         this.IpLoadBalancerHomeService = IpLoadBalancerHomeService;
         this.quota = Object.assign({}, quota);
 
@@ -12,19 +13,12 @@ class IpLoadBalancerUpdateQuotaCtrl {
     }
 
     updateQuota () {
-        this.saving = true;
-        this.IpLoadBalancerHomeService.updateQuota(
-            this.$stateParams.serviceName,
-            this.quota.zone,
-            // Convert GiB to bytes
-            this.alert * Math.pow(1000, 3)
-        ).then(result => {
-            this.$uibModalInstance.close(result);
-        }).catch(err => {
-            this.$uibModalInstance.close(err);
-        }).finally(() => {
-            this.saving = false;
+        this.update = this.ControllerHelper.request.getHashLoader({
+            loaderFunction: () => this.IpLoadBalancerHomeService.updateQuota(this.$stateParams.serviceName, this.quota.zone, this.alert * Math.pow(1000, 3))
+                .then(response => this.$uibModalInstance.close(response))
+                .catch(error => this.$uibModalInstance.close(error))
         });
+        return this.update.load();
     }
 
     cancel () {