+++ client/app/iplb/serverFarm/delete/iplb-server-farm-delete.controller.js
@@ -1,7 +1,8 @@
 class IpLoadBalancerServerFarmDeleteCtrl {
-    constructor ($stateParams, $uibModalInstance, farm, IpLoadBalancerServerFarmService) {
+    constructor ($stateParams, $uibModalInstance, ControllerHelper, farm, IpLoadBalancerServerFarmService) {
         this.$stateParams = $stateParams;
         this.$uibModalInstance = $uibModalInstance;
+        this.ControllerHelper = ControllerHelper;
         this.IpLoadBalancerServerFarmService = IpLoadBalancerServerFarmService;
 
         this.farm = farm;
@@ -11,13 +12,12 @@ class IpLoadBalancerServerFarmDeleteCtrl {
     }
 
     confirm () {
-        this.saving = true;
-        return this.IpLoadBalancerServerFarmService.delete(this.type, this.$stateParams.serviceName, this.farmId)
-            .then(response => this.$uibModalInstance.close(response))
-            .catch(response => this.$uibModalInstance.dismiss(response))
-            .finally(() => {
-                this.saving = false;
-            });
+        this.delete = this.ControllerHelper.request.getHashLoader({
+            loaderFunction: () => this.IpLoadBalancerServerFarmService.delete(this.type, this.$stateParams.serviceName, this.farmId)
+                .then(response => this.$uibModalInstance.close(response))
+                .catch(error => this.$uibModalInstance.dismiss(error))
+        });
+        return this.delete.load();
     }
 
     cancel () {