+++ client/app/ui-components/button/button.component.js
@@ -1,75 +1,43 @@
 angular.module("managerApp")
-    .component("cuiButton", {
-        template: `
-            <button class="oui-button oui-button_{{ $ctrl.type }}"
-                    type="submit"
-                    data-ng-bind="$ctrl.text"
-                    data-ng-click="$ctrl.onClick()"
-                    data-ng-disabled="$ctrl.disabled"></button>
-        `,
-        bindings: {
-            text: "<",
-            onClick: "&",
-            disabled: "<",
-            type: "@"
-        }
-    })
     .component("cuiSubmitButton", {
         template: `
-            <button class="oui-button oui-button_primary"
-                    type="submit"
-                    data-ng-bind="$ctrl.text"
-                    data-ng-click="$ctrl.onClick()"
-                data-ng-disabled="$ctrl.disabled"></button>
+            <oui-button data-variant="primary"
+                data-type="submit"
+                data-on-click="$ctrl.onClick()"
+                data-disabled="$ctrl.disabled"
+                data-text="{{::$ctrl.text}}"
+                data-aria-label="{{::$ctrl.ariaLabel}}"
+                data-id="{{::$ctrl.id}}"
+                data-name="{{::$ctrl.name}}">
+            </oui-button>
         `,
         bindings: {
             text: "<",
             onClick: "&",
-            disabled: "<"
+            disabled: "<?",
+            ariaLabel: "@?",
+            id: "@?",
+            name: "@?"
         }
     })
     .component("cuiCancelButton", {
         template: `
-            <button class="oui-button oui-button_secondary"
-                    type="button"
-                    data-ng-bind="$ctrl.text"
-                    data-ng-click="$ctrl.onClick()"
-                    data-ng-disabled="$ctrl.disabled"></button>
+            <oui-button data-variant="secondary"
+                data-type="button"
+                data-on-click="$ctrl.onClick()"
+                data-disabled="$ctrl.disabled"
+                data-text="{{::$ctrl.text}}"
+                data-aria-label="{{::$ctrl.ariaLabel}}"
+                data-id="{{::$ctrl.id}}"
+                data-name="{{::$ctrl.name}}">
+            </oui-button>
         `,
         bindings: {
             text: "<",
             onClick: "&",
-            disabled: "<"
-        }
-    })
-    .component("cuiActionButton", {
-        template: `
-            <cui-button data-ng-if="$ctrl.action.callback"
-                data-text="$ctrl.action.text" 
-                data-on-click="$ctrl.action.callback()"
-                data-disabled="$ctrl.action.isAvailable && !$ctrl.action.isAvailable()"
-                data-type="{{ $ctrl.type }}"></cui-button>
-            <a data-ng-if="$ctrl.action.state"
-                class="oui-button oui-button_{{ $ctrl.type }}"
-                data-ui-sref="{{ $ctrl.action.state + $ctrl.getActionStateParamString() }}"
-                data-ng-bind="$ctrl.action.text" 
-                data-ng-disabled="$ctrl.action.isAvailable && !$ctrl.action.isAvailable()"
-                data-type="{{ $ctrl.type }}"></a>
-            <a data-ng-if="$ctrl.action.href"
-                class="oui-button oui-button_{{ $ctrl.type }}"
-                data-ng-href="{{ $ctrl.action.href }}"
-                data-ng-bind="$ctrl.action.text"
-                data-ng-disabled="$ctrl.action.isAvailable && !$ctrl.action.isAvailable()"
-                data-type="{{ $ctrl.type }}"></a>
-        `,
-        controller: class cuiActionButtonController {
-            getActionStateParamString () {
-                if (!this.action.stateParams) { return ""; }
-                return `(${JSON.stringify(this.action.stateParams)})`;
-            }
-        },
-        bindings: {
-            action: "<",
-            type: "@"
+            disabled: "<?",
+            ariaLabel: "@?",
+            id: "@?",
+            name: "@?"
         }
     });